Module propfuzz::prelude

Re-exports for the most commonly used APIs of propfuzz.

This includes:

the propfuzz macro from this crate

the entire prelude of proptest, so existing tests can be migrated with minimal hassle.

Examples

There's no need to specify proptest as a separate dependency, since you can write:

use propfuzz::prelude::*;
use proptest::collection::vec;

/// Example test.
#[propfuzz]
fn test(#[propfuzz(strategy = "vec(any::<u8>(), 0..64)")] v: Vec<u8>) {
    // ...
}

Re-exports

pub use crate::propfuzz;

pub use proptest;

pub use proptest;

pub use proptest::prelude::*;