Trait propfuzz::traits::StructuredTarget

pub trait StructuredTarget: Send + Sync + Debug {
    type Value: Debug;
    fn name(&self) -> &'static str;

    fn description(&self) -> Option<&'static str>;

    fn execute(
        &self, 
        test_runner: &mut TestRunner
    ) -> Result<(), TestError<Self::Value>>;

    fn proptest_config(&self) -> ProptestConfig { ... }

    fn fmt_value(&self, value: &Self::Value, f: &mut Formatter) -> Result { ... }
}

Represents a structured fuzz target.

Functions written with the propfuzz macro are converted to implementations of this trait.

A trait that implements Propfuzz can be used both as a standard property-based test, and as a target for structured, mutation-based fuzzing.

Structured, mutation-based fuzzers use random byte sequences as a pass-through RNG. In other words, the random byte sequences act as something like a DNA for random values.

Associated Types

type Value: Debug

The type of values generated by the proptest.

Loading content...
Required methods

fn name(&self) -> &'static str

Returns the name of this structured fuzz target.

fn description(&self) -> Option<&'static str>

Returns an optional description for this structured fuzz target.

fn execute(
&self,
test_runner: &mut TestRunner
) -> Result<(), TestError<Self::Value>>

Executes this test using the given test runner.

This is where the main body of the test goes.

Loading content...
Provided methods

fn proptest_config(&self) -> ProptestConfig

Returns the proptest config for this fuzz target.

The default implementation for the #[propfuzz] macro uses the default proptest config, and sets source_file to the file the macro is invoked from. The config can be altered through passing arguments to #[propfuzz].

fn fmt_value(&self, value: &Self::Value, f: &mut Formatter) -> Result

Formats a failing test case for displaying to the user.

The default implementation calls the fmt::Debug implementation on the value.

The default implementation for the #[propfuzz] macro prints out variable names and values.

Loading content...